@@ -1,10 +1,14 @@
 {
   "devDependencies": {
-    "lerna": "5.1.8",
-    "prettier": "2.7.1",
-    "rimraf": "3.0.2",
-    "ts-node": "10.9.1",
-    "typescript": "4.7.4"
+    "@tstv/eslint-config": "2.0.1",
+    "eslint": "^8.45.0",
+    "lefthook": "^1.4.6",
+    "lerna": "^7.1.4",
+    "prettier": "^2.7.1",
+    "pretty-quick": "^3.1.3",
+    "rimraf": "^5.0.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.1.6"
   },
   "name": "ts-lerna-template",
   "private": true,
@@ -12,7 +16,7 @@
     "boot": "lerna bootstrap --reject-cycles",
     "clean": "lerna run clean --parallel",
     "dist": "lerna run dist",
-    "release": "yarn install && yarn dist && lerna publish && yarn clean",
+    "release": "npm run install && npm run dist && lerna publish && npm run clean",
     "style": "prettier --ignore-path .gitignore \"**/*.{md,json,html,ts}\" --write",
     "test": "lerna run --no-sort --concurrency 8 test"
   },
