@@ -1,25 +1,24 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import {PriceCalculator} from '@tstv/business';
+import React from "react";
+import ReactDOM from "react-dom/client";
+import { PriceCalculator } from "@tstv/business";
 
 // https://typescript.tv/react/use-react-with-typescript/
 
-interface DivWrapperProps extends React.HTMLProps<HTMLDivElement> {
-}
+interface DivWrapperProps extends React.HTMLProps<HTMLDivElement> {}
 
-const DivWrapper: React.FC<DivWrapperProps> = (props) => (
-  <div {...props} />
-);
+const DivWrapper: React.FC<DivWrapperProps> = (props) => <div {...props} />;
 
-interface AppProps {
-}
+interface AppProps {}
 
 const App: React.FC<AppProps> = () => (
   <>
     <h1>Hello, {PriceCalculator.getTotal(100)}!</h1>
-    <DivWrapper children={<span>Hello, World!</span>}/>
+    <DivWrapper children={<span>Hello, World!</span>} />
   </>
 );
-const rootNode = document.getElementById('root');
+const rootNode = document.getElementById("root");
 
-ReactDOM.render(<App/>, rootNode);
+if (rootNode) {
+  const root = ReactDOM.createRoot(rootNode);
+  root.render(<App />);
+}
